+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Application/ChargeLinks/Handlers/ChargeLinksReceivedEventHandler.cs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using GreenEnergyHub.Charges.Domain.ChargeLinks;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeLinksAcceptedEvents;
@@ -40,7 +39,7 @@ namespace GreenEnergyHub.Charges.Application.ChargeLinks.Handlers
             _chargeLinkRepository = chargeLinkRepository;
         }
 
-        public async Task HandleAsync([NotNull] ChargeLinksReceivedEvent chargeLinksReceivedEvent)
+        public async Task HandleAsync(ChargeLinksReceivedEvent chargeLinksReceivedEvent)
         {
             // Upcoming stories will cover the update scenarios where charge link already exists
             var chargeLinks = await _chargeLinkFactory.CreateAsync(chargeLinksReceivedEvent).ConfigureAwait(false);