+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Application/ChargeLinks/Handlers/CreateLinkRequestHandler.cs
@@ -22,7 +22,7 @@ using GreenEnergyHub.Charges.Contracts;
 using GreenEnergyHub.Charges.Domain.DefaultChargeLinks;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeLinksCommands;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeLinksReceivedEvents;
-using GreenEnergyHub.Charges.Domain.Dtos.CreateLinksRequests;
+using GreenEnergyHub.Charges.Domain.Dtos.CreateDefaultChargeLinksRequests;
 using GreenEnergyHub.Charges.Domain.MeteringPoints;
 using GreenEnergyHub.Charges.Infrastructure.Core.Correlation;
 using GreenEnergyHub.Charges.Infrastructure.Core.MessageMetaData;
@@ -66,22 +66,22 @@ namespace GreenEnergyHub.Charges.Application.ChargeLinks.Handlers
             _logger = loggerFactory.CreateLogger(nameof(CreateLinkRequestHandler));
         }
 
-        public async Task HandleAsync([NotNull] CreateLinksRequest createLinksRequest)
+        public async Task HandleAsync(CreateDefaultChargeLinksRequest createDefaultChargeLinksRequest)
         {
             if (!_messageMetaDataContext.IsReplyToSet())
             {
-                _logger.LogError($"The reply queue name was absent or empty, could not handle request CreateDefaultChargeLinks on metering point with id: {createLinksRequest.MeteringPointId} and correlation id: {_correlationIdContext.Id}");
+                _logger.LogError($"The reply queue name was absent or empty, could not handle request CreateDefaultChargeLinks on metering point with id: {createDefaultChargeLinksRequest.MeteringPointId} and correlation id: {_correlationIdContext.Id}");
 
                 throw new ArgumentNullException(nameof(_messageMetaDataContext.ReplyTo));
             }
 
             var meteringPoint =
-                await GetMeteringPointAsync(createLinksRequest.MeteringPointId).ConfigureAwait(false);
+                await GetMeteringPointAsync(createDefaultChargeLinksRequest.MeteringPointId).ConfigureAwait(false);
 
             if (meteringPoint == null)
             {
                 await ReplyWithFailedAsync(
-                    createLinksRequest.MeteringPointId,
+                    createDefaultChargeLinksRequest.MeteringPointId,
                     _messageMetaDataContext.ReplyTo).ConfigureAwait(false);
                 return;
             }
@@ -98,7 +98,7 @@ namespace GreenEnergyHub.Charges.Application.ChargeLinks.Handlers
             }
 
             await CreateAndDispatchChargeLinksReceivedEventIfApplicableForLinkingAsync(
-                createLinksRequest,
+                createDefaultChargeLinksRequest,
                 defaultChargeLinks,
                 meteringPoint);
         }
@@ -139,7 +139,7 @@ namespace GreenEnergyHub.Charges.Application.ChargeLinks.Handlers
         }
 
         private async Task CreateAndDispatchChargeLinksReceivedEventIfApplicableForLinkingAsync(
-            CreateLinksRequest createLinksRequest,
+            CreateDefaultChargeLinksRequest createDefaultChargeLinksRequest,
             IEnumerable<DefaultChargeLink> defaultChargeLinks,
             MeteringPoint meteringPoint)
         {
@@ -149,7 +149,7 @@ namespace GreenEnergyHub.Charges.Application.ChargeLinks.Handlers
                     meteringPoint.MeteringPointType)).ToList();
 
             var chargeLinksCommand = await _chargeLinksCommandFactory.CreateAsync(
-                createLinksRequest, chargeLinksApplicableForLinking);
+                createDefaultChargeLinksRequest, chargeLinksApplicableForLinking);
 
             var chargeLinksReceivedEvent = new ChargeLinksReceivedEvent(
                 _clock.GetCurrentInstant(),