+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/Dtos/ChargeLinksCommands/IChargeLinksCommandFactory.cs
@@ -16,14 +16,14 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using GreenEnergyHub.Charges.Domain.DefaultChargeLinks;
-using GreenEnergyHub.Charges.Domain.Dtos.CreateLinksRequests;
+using GreenEnergyHub.Charges.Domain.Dtos.CreateDefaultChargeLinksRequests;
 
 namespace GreenEnergyHub.Charges.Domain.Dtos.ChargeLinksCommands
 {
     public interface IChargeLinksCommandFactory
     {
         Task<ChargeLinksCommand> CreateAsync(
-            [NotNull] CreateLinksRequest createLinksRequest,
+            [NotNull] CreateDefaultChargeLinksRequest createDefaultChargeLinksRequest,
             [NotNull] IReadOnlyCollection<DefaultChargeLink> defaultChargeLinks);
     }
 }