+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients/documents/release-notes/release-notes.md
@@ -1,5 +1,13 @@
 # Energinet.DataHub.Charges.Clients Release notes
 
+## Version 1.0.21
+
+Revert to using record for `ChargeLinkDto`.
+
+## Version 1.0.20
+
+Use class for `ChargeLinkDto` instead of record.
+
 ## Version 1.0.19
 
 `IChargeLinkClient` returns charge links with new naming.